Move webview shared code into platform
