+++ src/platform/webviewHost/webviewHost.ts
@@ -3,7 +3,7 @@
 
 'use strict';
 
-import '../../../platform/common/extensions';
+import '../common/extensions';
 
 import {
     ConfigurationChangeEvent,
@@ -14,19 +14,18 @@ import {
     WebviewView as vscodeWebviewView,
     WorkspaceConfiguration
 } from 'vscode';
-import { IWebview, IWorkspaceService } from '../../../platform/common/application/types';
-import { isTestExecution } from '../../../platform/common/constants';
-import { traceInfo } from '../../../platform/logging';
-import { Resource, IConfigurationService, IDisposable } from '../../../platform/common/types';
-import { Deferred, createDeferred } from '../../../platform/common/utils/async';
-import { testOnlyMethod } from '../../../platform/common/utils/decorators';
-import * as localize from '../../../platform/common/utils/localize';
-import { StopWatch } from '../../../platform/common/utils/stopWatch';
-import { InteractiveWindowMessages, SharedMessages } from '../../../platform/messageTypes';
-import { sendTelemetryEvent } from '../../../telemetry';
-import { DefaultTheme, PythonExtension, Telemetry } from '../../webview-side/common/constants';
-import { IJupyterExtraSettings } from '../types';
-import { getOSType, OSType } from '../../../platform/common/utils/platform';
+import { IWebview, IWorkspaceService } from '../common/application/types';
+import { DefaultTheme, isTestExecution, PythonExtension, Telemetry } from '../common/constants';
+import { traceInfo } from '../logging';
+import { Resource, IConfigurationService, IDisposable } from '../common/types';
+import { Deferred, createDeferred } from '../common/utils/async';
+import { testOnlyMethod } from '../common/utils/decorators';
+import * as localize from '../common/utils/localize';
+import { StopWatch } from '../common/utils/stopWatch';
+import { InteractiveWindowMessages, SharedMessages } from '../messageTypes';
+import { sendTelemetryEvent } from '../../telemetry';
+import { IJupyterExtraSettings } from './types';
+import { getOSType, OSType } from '../common/utils/platform';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
