+++ src/platform/webviewHost/webviewPanelProvider.ts
@@ -4,10 +4,10 @@
 
 import { inject, injectable } from 'inversify';
 import { Uri, ViewColumn, WebviewOptions, WebviewPanel as vscodeWebviewPanel, window } from 'vscode';
-import { IWebviewPanel, IWebviewPanelOptions, IWebviewPanelProvider } from '../../../platform/common/application/types';
-import { IFileSystem } from '../../../platform/common/platform/types';
-import { IDisposableRegistry, IExtensionContext } from '../../../platform/common/types';
-import * as path from '../../../platform/vscode-path/path';
+import { IWebviewPanel, IWebviewPanelOptions, IWebviewPanelProvider } from '../common/application/types';
+import { IFileSystem } from '../common/platform/types';
+import { IDisposableRegistry, IExtensionContext } from '../common/types';
+import * as path from '../vscode-path/path';
 import { Webview } from './webview';
 
 class WebviewPanel extends Webview implements IWebviewPanel {
