+++ src/webviews/webview-side/interactive-common/mainState.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 'use strict';
-import { IJupyterExtraSettings } from '../../extension-side/types';
+import { IJupyterExtraSettings } from '../../../platform/webviewHost/types';
 import { getDefaultSettings } from '../react-common/settingsReactSide';
 
 export enum CursorPos {
